# doc/api/swagger.yaml
basePath: /api/v2
definitions:
...
  database.Scenario:
    properties:
      id:
        type: integer
      name:
        description: Name of scenario
        type: string
      running:
        description: Running state of scenario
        type: boolean
      startParameters:
        description: Start parameters of scenario as JSON
        type: string
    type: object
...
  database.User:
    properties:
      active:
        description: Indicating status of user (false means user is inactive and should
          not be able to login)
        type: boolean
      id:
        type: integer
      mail:
        description: Mail of user
        type: string
      role:
        description: Role of user
        type: string
      username:
        description: Username of user
        type: string
    type: object
...
  docs.ResponseError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  docs.ResponseScenario:
    properties:
      scenario:
        $ref: '#/definitions/database.Scenario'
        type: object
    type: object
  docs.ResponseScenarios:
    properties:
      scenarios:
        items:
          $ref: '#/definitions/database.Scenario'
        type: array
    type: object
...
  docs.ResponseUser:
    properties:
      user:
        $ref: '#/definitions/database.User'
        type: object
    type: object
  docs.ResponseUsers:
    properties:
      users:
        items:
          $ref: '#/definitions/database.User'
        type: array
    type: object
...
  scenario.addScenarioRequest:
    properties:
      scenario:
        $ref: '#/definitions/scenario.validNewScenario'
        type: object
    type: object
  scenario.updateScenarioRequest:
    properties:
      scenario:
        $ref: '#/definitions/scenario.validUpdatedScenario'
        type: object
    type: object
  scenario.validNewScenario:
    properties:
      Name:
        type: string
      Running:
        type: boolean
      StartParameters:
        type: string
    required:
    - Name
    - StartParameters
    type: object
  scenario.validUpdatedScenario:
    properties:
      Name:
        type: string
      Running:
        type: boolean
      StartParameters:
        type: string
    type: object
...
  user.addUserRequest:
    properties:
      user:
        $ref: '#/definitions/user.validNewUser'
        type: object
    type: object
...
  user.updateUserRequest:
    properties:
      user:
        $ref: '#/definitions/user.validUpdatedRequest'
        type: object
    type: object
  user.validNewUser:
    properties:
      Mail:
        type: string
      Password:
        type: string
      Role:
        type: string
      Username:
        type: string
    required:
    - Mail
    - Password
    - Role
    - Username
    type: object
  user.validUpdatedRequest:
    properties:
      Active:
        type: string
      Mail:
        type: string
      OldPassword:
        type: string
      Password:
        type: string
      Role:
        type: string
      Username:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: sonja.happ@eonerc.rwth-aachen.de
    name: Sonja Happ
  description: |-
    This is the VILLASweb Backend API v2.0.
    Parts of this API are still in development. Please check https://git.rwth-aachen.de/acs/public/villas/web-backend-go for more information.
  license:
    name: GNU GPL 3.0
    url: http://www.gnu.de/documents/gpl-3.0.en.html
  title: VILLASweb Backend API
  version: "2.0"
paths:
...
  /scenarios:
    get:
      operationId: getScenarios
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scenarios to which user has access
          schema:
            $ref: '#/definitions/docs.ResponseScenarios'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all scenarios of requesting user
      tags:
      - scenarios
    post:
      consumes:
      - application/json
      operationId: addScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario to be added
        in: body
        name: inputScenario
        required: true
        schema:
          $ref: '#/definitions/scenario.addScenarioRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: scenario that was added
          schema:
            $ref: '#/definitions/docs.ResponseScenario'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a scenario
      tags:
      - scenarios
  /scenarios/{scenarioID}:
    delete:
      operationId: deleteScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted scenario
          schema:
            $ref: '#/definitions/docs.ResponseScenario'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a scenario
      tags:
      - scenarios
    get:
      operationId: getScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Scenario requested by user
          schema:
            $ref: '#/definitions/docs.ResponseScenario'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get scenario
      tags:
      - scenarios
    put:
      consumes:
      - application/json
      operationId: updateScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario to be updated
        in: body
        name: inputScenario
        required: true
        schema:
          $ref: '#/definitions/scenario.updateScenarioRequest'
          type: object
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated scenario.
          schema:
            $ref: '#/definitions/docs.ResponseScenario'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Update a scenario
      tags:
      - scenarios
  /scenarios/{scenarioID}/user:
    delete:
      operationId: deleteUserFromScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      - description: User name
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User that was deleted from scenario
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a user from a scenario
      tags:
      - scenarios
    put:
      operationId: addUserToScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      - description: User name
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User that was added to scenario
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a user to a a scenario
      tags:
      - scenarios
  /scenarios/{scenarioID}/users/:
    get:
      operationId: getUsersOfScenario
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Scenario ID
        in: path
        name: scenarioID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array of users that have access to the scenario
          schema:
            $ref: '#/definitions/docs.ResponseUsers'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get users of a scenario
      tags:
      - scenarios
...
  /users:
    get:
      operationId: GetUsers
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of users
          schema:
            $ref: '#/definitions/docs.ResponseUsers'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      operationId: AddUser
      parameters:
      - description: User to be added
        in: body
        name: inputUser
        required: true
        schema:
          $ref: '#/definitions/user.addUserRequest'
          type: object
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Contains added user object
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a user
      tags:
      - users
  /users/{userID}:
    delete:
      operationId: DeleteUser
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted user
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      operationId: GetUser
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: requested user
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "403":
          description: Access forbidden.
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      operationId: UpdateUser
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User to be updated (anything except for ID can be changed, role
          can only be change by admin)
        in: body
        name: inputUser
        required: true
        schema:
          $ref: '#/definitions/user.updateUserRequest'
          type: object
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Contains updated user
          schema:
            $ref: '#/definitions/docs.ResponseUser'
            type: object
        "400":
          description: Bad request.
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "403":
          description: Access forbidden.
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Update a user
      tags:
      - users
...
swagger: "2.0"